administration/api/serializers/article.py
from rest_framework import serializers

from administration.models.article import Article


class ArticleSerializer(serializers.ModelSerializer):
    class Meta:
        model = Article
        fields = ('title', 'url', 'summary')
